+++ spec/active_record/materialized/metadata_spec.rb
@@ -3,13 +3,7 @@
 require "spec_helper"
 
 RSpec.describe ActiveRecord::Materialized::Metadata do
-  let(:view_class) do
-    Class.new(ActiveRecord::Materialized::View) do
-      self.table_name = "mv_metadata_test"
-      materialized_from { ViewSources.item_id_sample }
-    end
-  end
-
+  let(:view_class) { define_view("mv_metadata_test", :item_id_sample) }
   let(:metadata) { described_class.new(view_class) }
 
   it "marks failed refreshes" do